Fixes https://github.com/nemtech/symbol-sdk-java/issues/310

+++ sdk-core/src/main/java/io/nem/symbol/sdk/api/BinarySerialization.java
@@ -60,6 +60,6 @@ public interface BinarySerialization {
      * @param transaction the transaction
      * @return the size of the transaction.
      */
-    <T extends Transaction> int getSize(T transaction);
+    <T extends Transaction> long getSize(T transaction);
 
 }
